% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{annual_loss}
\alias{annual_loss}
\title{Aggregate annual loss by year}
\usage{
annual_loss(x)
}
\arguments{
\item{x}{a YLT}
}
\value{
a data.table with rows = sims
}
\description{
Aggregate annual loss by year
}
