% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catr.R
\docType{package}
\name{catr}
\alias{catr}
\title{catr: a package with functions to help analyse Catastrophe model data}
\description{
catr provides functions to help
}
\section{catr functions}{

The catr functions...
}

