% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{create_elt}
\alias{create_elt}
\title{Create an Event Loss Table with parameters for simulation}
\usage{
create_elt(dt)
}
\arguments{
\item{dt}{a data.table}
}
\value{
ELT
}
\description{
Create an Event Loss Table with parameters for simulation
}
