% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{create_ep_curve}
\alias{create_ep_curve}
\title{Calculate Occurrence Exceedance Probability}
\usage{
create_ep_curve(x, y, z)
}
\arguments{
\item{x}{a annual loss table}

\item{y}{a years}

\item{z}{the loss amount}
}
\value{
ep curve
}
\description{
Calculate Occurrence Exceedance Probability
}
