% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{create_oep_curve}
\alias{create_oep_curve}
\title{OEP Curve}
\usage{
create_oep_curve(
  dt,
  y,
  z,
  rp = c(10000, 5000, 1000, 500, 250, 200, 100, 50, 25, 10, 5, 2)
)
}
\arguments{
\item{dt}{aggregate annual YLT}

\item{y}{vector of year}

\item{z}{vector of loss amount}

\item{rp}{return period default points=  c(10000,5000,1000,500,250,200,100,50, 25,10,5 , 2)}
}
\value{
a vector of OEP at return periods as specified by the argument rp
}
\description{
OEP Curve
}
\examples{
create_oep_curve(data.table::data.table("Year" = c(1,2,3,4,5) , 
          "Loss" =c(1 , 20 , 500 , 100 , 10000)) , y= "Year", z="Loss")
}
