% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{create_ylt}
\alias{create_ylt}
\title{TCreate a YLT from ELT}
\usage{
create_ylt(dt, sims, ann_rate, event_id, expval, mu)
}
\arguments{
\item{mu}{}
}
\value{

}
\description{
TCreate a YLT from ELT
}
