% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eltr.R
\docType{package}
\name{eltr}
\alias{eltr}
\title{eltr: a package with functions to help analyse Catastrophe model data}
\description{
eltr provides functions to help
}
\section{eltr functions}{

The eltr functions...
}

