% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{layer_loss}
\alias{layer_loss}
\title{Limited loss to the layer}
\usage{
layer_loss(x, Excess, Limit)
}
\arguments{
\item{x}{event loss}

\item{Excess}{treaty retention}

\item{Limit}{treaty limit}
}
\value{
limited loss to the layer
}
\description{
Limited loss to the layer
}
\examples{
layer_loss(5,2,6)
layer_loss(5,10,6)

}
