% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{create_ylt}
\alias{create_ylt}
\title{Create a Year loss table in tidy format}
\usage{
create_ylt(x, sims)
}
\arguments{
\item{x}{a data.table of elt  with parameters}

\item{sims}{number of years to simulate}
}
\value{
YLT
}
\description{
Create a Year loss table in tidy format
}
